refactor: 음성 녹음 화면 구조 개선 및 비즈니스 로직 리팩터링
- 음성 녹음 화면(`VoiceRecordingScreen.kt`)에서 대규모 UI 컴포넌트들을 별도 파일로 분리
    - `VoiceRecordingButtonArea.kt`: 녹음 상태별 버튼 영역 분리
    - `VoiceRecordingContent.kt`: 스크립트 및 녹음 상태 표시 영역 분리
    - `VoiceRecordingStateProperties.kt`: `AudioSessionState` 기반 확장 프로퍼티 정리
    - `VoiceRecordingStatusBarStyle.kt`: 상태바 스타일 제어 로직 분리
- `AnalysisFlowViewModel` 내 UI 상태 변경 및 분석 요청 로직 개선
    - `reduceFormOrNull` 확장 함수를 통해 `AnalysisFlowUiIntent`에 따른 `AnalysisForm` 갱신 로직 간소화
    - `analyzePresentationRecording` 로직을 별도 함수로 추출하고 파일 캐시 처리 방식 개선
    - 스크립트 건너뛰기(`skipScript`) 및 파일 업로드 재시도 로직 보완
- `MediaRecordingAudioController` 내 `AudioSessionEffect` 전송 방식을 `emit` 확장 함수로 통일
- `ScriptInputScreen` 내 직접 입력 모드에서 텍스트 필드 하단 여백(`Spacer`) 추가 및 미리보기 케이스 추가
- `AnalysisScreen` 내 스낵바 표시 및 녹음 권한 핸들러 호출 로직 구조 개선 (컴포넌트 분리)
